+++ src/main/java/yapp/buddycon/app/gifticon/adapter/infra/jpa/GifticonJpaRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import yapp.buddycon.app.gifticon.adapter.client.response.GifticonResponseDTO;
-import yapp.buddycon.app.gifticon.adapter.infra.entity.GifticonEntity;
 import yapp.buddycon.app.gifticon.domain.GifticonStoreCategory;
 
 import java.util.Optional;
